simplify the BSP to keep are timer_f and include the RTC library
Include the STM32RTC library to all the BSP as only few functions
remain in the BSP as the RTC is instantiated by the STM32RTC library
