+++ src/BSP/timer_if.h
@@ -46,6 +46,7 @@ void UTIL_TIMER_IRQ_MAP_PROCESS(void *data);
 /* Exported types ------------------------------------------------------------*/
 /* USER CODE BEGIN ET */
 
+
 /* USER CODE END ET */
 
 /* Exported constants --------------------------------------------------------*/
